+++ docs/11_0_0/gettingstarted/whatsnew.md
@@ -4,7 +4,8 @@ This page contains a list of big features. Please check the GitHub issues for al
 
 ## 11.0.0
 
-  * TODO
+  * AutoComplete: added `footer` facet which will be added to the suggestion list.
+  * InputText, InputTextarea: counter can count bytes instead of characters.
 
 Look into [migration guide](https://primefaces.github.io/primefaces/11_0_0/#/../migrationguide/10_0_0?id=datatable) for more enhancements and changes.
-Or check the list of 600+ issues closed for [11.0.0](https://github.com/primefaces/primefaces/issues?q=is%3Aclosed+milestone%3A11.0.0), 
+Or check the list of 600+ issues closed for [11.0.0](https://github.com/primefaces/primefaces/issues?q=is%3Aclosed+milestone%3A11.0.0),